+++ src/tiled/objectsview.cpp
@@ -35,6 +35,7 @@
 #include <QKeyEvent>
 #include <QMenu>
 #include <QPainter>
+#include <QScopedValueRollback>
 
 namespace Tiled {
 
@@ -79,8 +80,10 @@ void ObjectsView::setMapDocument(MapDocument *mapDoc)
     if (mapDoc == mMapDocument)
         return;
 
-    if (mMapDocument)
+    if (mMapDocument) {
+        saveExpandedLayers();
         mMapDocument->disconnect(this);
+    }
 
     mMapDocument = mapDoc;
 
@@ -100,6 +103,8 @@ void ObjectsView::setMapDocument(MapDocument *mapDoc)
 
         if (mActiveFilter)
             expandAll();
+        else
+            restoreExpandedLayers();
     } else {
         mProxyModel->setSourceModel(nullptr);
     }
@@ -150,7 +155,7 @@ void ObjectsView::saveExpandedLayers()
     if (mActiveFilter)
         return;
 
-    mExpandedLayers[mMapDocument].clear();
+    mMapDocument->expandedObjectLayers.clear();
 
     for (Layer *layer : mMapDocument->map()->allLayers()) {
         if (!layer->isObjectGroup() && !layer->isGroupLayer())
@@ -159,16 +164,13 @@ void ObjectsView::saveExpandedLayers()
         const QModelIndex sourceIndex = mMapDocument->mapObjectModel()->index(layer);
         const QModelIndex index = mProxyModel->mapFromSource(sourceIndex);
         if (isExpanded(index))
-            mExpandedLayers[mMapDocument].append(layer->id());
+            mMapDocument->expandedObjectLayers.insert(layer->id());
     }
 }
 
 void ObjectsView::restoreExpandedLayers()
 {
-    if (mActiveFilter)
-        return;
-
-    const auto layers = mExpandedLayers.take(mMapDocument);
+    const auto &layers = mMapDocument->expandedObjectLayers;
     for (const int layerId : layers) {
         if (Layer *layer = mMapDocument->map()->findLayerById(layerId)) {
             const QModelIndex sourceIndex = mMapDocument->mapObjectModel()->index(layer);
@@ -176,11 +178,6 @@ void ObjectsView::restoreExpandedLayers()
             setExpanded(index, true);
         }
     }
-}
-
-void ObjectsView::clearExpandedLayers(MapDocument *mapDocument)
-{
-    mExpandedLayers.remove(mapDocument);
 }
 
 bool ObjectsView::event(QEvent *event)
@@ -275,9 +272,8 @@ void ObjectsView::selectionChanged(const QItemSelection &selected,
     }
 
     if (selectedObjects != mMapDocument->selectedObjects()) {
-        mSynching = true;
+        QScopedValueRollback<bool> synching(mSynching, true);
         mMapDocument->setSelectedObjects(selectedObjects);
-        mSynching = false;
     }
 }
 
@@ -374,12 +370,11 @@ void ObjectsView::synchronizeSelectedItems()
         itemSelection.select(index, index);
     }
 
-    mSynching = true;
+    QScopedValueRollback<bool> synching(mSynching, true);
     selectionModel()->select(itemSelection,
                              QItemSelectionModel::Select |
                              QItemSelectionModel::Rows |
                              QItemSelectionModel::Clear);
-    mSynching = false;
 }
 
 void ObjectsView::expandToSelectedObjects()