+++ src/tiled/session.h
@@ -113,6 +113,26 @@ inline QVariant toSettingsValue<QPointF>(const QPointF &point)
     };
 }
 
+template<>
+inline QSet<int> fromSettingsValue<QSet<int>>(const QVariant &value)
+{
+    const auto variantList = value.toList();
+    QSet<int> set;
+    for (const auto &variantValue : variantList)
+        set.insert(variantValue.value<int>());
+    return set;
+}
+
+template<>
+inline QVariant toSettingsValue<QSet<int>>(const QSet<int> &set)
+{
+    QVariantList variantList;
+    variantList.reserve(set.size());
+    for (const int value : set)
+        variantList.append(value);
+    return variantList;
+}
+
 
 class Session : protected FileHelper
 {